Restructure to make reuse easier
